pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "xhs-likes-manager"
version = "0.1.0"
description = "Manage your Xiaohongshu (小红书) likes and bookmarks locally"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40",
    "pyyaml>=6.0",
]

[project.scripts]
xhs-likes-manager = "xhs_likes_manager.cli:main"

[tool.setuptools.packages.find]
include = ["xhs_likes_manager*"]
